backend/requirements.txt
fastapi
uvicorn
alembic
sqlalchemy
psycopg2-binary
pydantic-settings
python-dotenv
black
beautifulsoup4
httpx
apscheduler~=3.10.0
httpx[http2]